make some auto_parallel uts deprecated
