+++ test/deprecated/auto_parallel/pipeline_scheduler_vpp_unittest_deprecated.py
