+++ test/deprecated/auto_parallel/pipeline_scheduler_zb_unittest.py
@@ -13,9 +13,12 @@
 # limitations under the License.
 
 import random
+import sys
 import unittest
 
 import numpy as np
+
+sys.path.append("../../auto_parallel")
 from get_gpt_model import FakeDataset, generate_model
 
 import paddle
