+++ test/deprecated/auto_parallel/random_control_unittest_deprecated.py
@@ -13,9 +13,13 @@
 # limitations under the License.
 
 import random
+import sys
 import unittest
 
 import numpy as np
+
+sys.path.append("../../auto_parallel")
+
 from get_gpt_model import FakeDataset, generate_model
 
 import paddle
