+++ test/deprecated/auto_parallel/test_pattern_deprecated.py
@@ -13,8 +13,10 @@
 # limitations under the License.
 
 
+import sys
 import unittest
 
+sys.path.append("../..")
 import auto_parallel_gpt_model as modeling
 import numpy as np
 from auto_parallel_gpt_model import (
@@ -93,7 +95,7 @@ def get_gpt_model(
     return train_program, start_program, loss, gen_data
 
 
-class TestPatternMatch(unittest.TestCase):
+class TestGroupOperatorsAndPatterns(unittest.TestCase):
     def test_gpt(self):
         modeling.init_global()
         train_program = static.Program()
@@ -110,19 +112,28 @@ class TestPatternMatch(unittest.TestCase):
             sequence_len,
             vocab_size,
         )
-        from paddle.distributed.auto_parallel.static.dist_context import (
-            DistributedContext,
-        )
         from paddle.distributed.auto_parallel.static.tuner.rule_based_tuner import (
+            _PATTERNS,
             GraphUtil,
-            RuleBasedTuner,
         )
 
-        dist_context = DistributedContext()
-        tuner = RuleBasedTuner(dist_context)
         graph = GraphUtil.convert_to_graph(train_program.global_block())
-        results = GraphUtil.match_all_patterns(graph)
-        print(results)
+        print("graph: ", graph)
+        print("qkv: ", _PATTERNS["qkv"].attrs["shard_spec"])
+        print("row_matmul: ", _PATTERNS["row_matmul"].attrs["shard_spec"])
+        print("ffn: ", _PATTERNS["ffn"].attrs["shard_spec"])
+        print(
+            "shared_word_embedding: ",
+            _PATTERNS["shared_word_embedding"].attrs["shard_spec"],
+        )
+        print(
+            "position_embedding: ",
+            _PATTERNS["position_embedding"].attrs["shard_spec"],
+        )
+        print(
+            "unsqueeze_data: ", _PATTERNS["unsqueeze_data"].attrs["shard_spec"]
+        )
+        print("reshape_data: ", _PATTERNS["reshape_data"].attrs["shard_spec"])
 
 
 if __name__ == "__main__":
