+++ test/deprecated/auto_parallel/test_pipeline_scheduler_zb_deprecated.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2022 PaddlePaddle Authors. All Rights Reserved.
+# Copyright (c) 2024 PaddlePaddle Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,10 +19,12 @@ import tempfile
 import unittest
 
 
-class TestRandomCtrlPass(unittest.TestCase):
-    def test_mp2_with_recompute(self):
+class TestZBH1Pass(unittest.TestCase):
+    def test_pp2(self):
         file_dir = os.path.dirname(os.path.abspath(__file__))
-        launch_model_path = os.path.join(file_dir, "random_control_unittest.py")
+        launch_model_path = os.path.join(
+            file_dir, "pipeline_scheduler_zb_unittest_deprecated.py"
+        )
 
         if os.environ.get("WITH_COVERAGE", "OFF") == "ON":
             coverage_args = ["-m", "coverage", "run", "--branch", "-p"]
