+++ test/deprecated/auto_parallel/test_relaunch_with_planner_deprecated.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
+# Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import json
 import os
 import subprocess
 import sys
@@ -19,11 +20,34 @@ import tempfile
 import unittest
 
 
-class TestVPPPass(unittest.TestCase):
-    def test_pp2(self):
+class TestPlannerReLaunch(unittest.TestCase):
+    def setUp(self):
+        self.temp_dir = tempfile.TemporaryDirectory()
+
+    def tearDown(self):
+        self.temp_dir.cleanup()
+
+    def test_relaunch_with_planner(self):
+        from test_auto_parallel_relaunch import cluster_json, mapping_json
+
+        cluster_json_path = os.path.join(
+            self.temp_dir.name, "auto_parallel_cluster.json"
+        )
+        mapping_json_path = os.path.join(
+            self.temp_dir.name, "auto_parallel_rank_mapping.json"
+        )
+
+        cluster_json_object = json.loads(cluster_json)
+        with open(cluster_json_path, "w") as cluster_json_file:
+            json.dump(cluster_json_object, cluster_json_file)
+
+        mapping_json_object = json.loads(mapping_json)
+        with open(mapping_json_path, "w") as mapping_json_file:
+            json.dump(mapping_json_object, mapping_json_file)
+
         file_dir = os.path.dirname(os.path.abspath(__file__))
         launch_model_path = os.path.join(
-            file_dir, "pipeline_scheduler_vpp_unittest.py"
+            file_dir, "auto_parallel_relaunch_with_planner_deprecated.py"
         )
 
         if os.environ.get("WITH_COVERAGE", "OFF") == "ON":
@@ -31,25 +55,25 @@ class TestVPPPass(unittest.TestCase):
         else:
             coverage_args = []
 
-        tmp_dir = tempfile.TemporaryDirectory()
         cmd = [
             sys.executable,
             "-u",
             *coverage_args,
             "-m",
             "paddle.distributed.launch",
-            "--devices",
-            "0,1",
             "--log_dir",
-            tmp_dir.name,
+            self.temp_dir.name,
+            "--cluster_topo_path",
+            cluster_json_path,
+            "--rank_mapping_path",
+            mapping_json_path,
+            "--enable_auto_mapping",
+            "True",
             launch_model_path,
         ]
-
         process = subprocess.Popen(cmd)
         process.wait()
         self.assertEqual(process.returncode, 0)
-
-        tmp_dir.cleanup()
 
 
 if __name__ == "__main__":
