+++ test/deprecated/auto_parallel/test_while_op_completion_deprecated.py
