+++ test/deprecated/auto_parallel/test_while_op_partition_deprecated.py
